Add styling details & js classes modules
