+++ index.html
@@ -18,15 +18,14 @@
     <h1>Awesome Books</h1>
     <section id="booksDisplay"><ul></ul></section>
     <form action="#" id="bookForm">
-      <br />
+      <br>
       <h2>Add a new Book</h2>
-      <input type="text" placeholder="Book title" required maxlength="20" />
-      <br />
-      <br />
-      <input type="text" placeholder="Author" required maxlength="10" /> <br />
-      <br />
+      <input type="text" placeholder="Book title" required maxlength="30">
+      <br>
+      <input type="text" placeholder="Author" required maxlength="20"> 
+      <br>
       <button type="submit" id="submitBttn">Add</button>
     </form>
-    <script src="./js/main.js"></script>
+    <script src="./js/main.js" type="module"></script>
   </body>
 </html>
