+++ js/main.js
@@ -1,96 +1,8 @@
-const booksDisplaySection = document.querySelector('#booksDisplay');
+import Book from './modules/Book.js';
+import Books from './modules/Books.js';
+import Methods from './modules/Methods.js';
+
 const form = document.querySelector('#submitBttn');
-
-class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
-
-class Books {
-  static ids = 0;
-
-  static booksData = [];
-
-  constructor() {
-    this.booksData = [];
-  }
-
-  static initializeBooks(data) {
-    this.booksData = data;
-  }
-
-  static getBooks() {
-    return this.booksData;
-  }
-}
-
-class Methods {
-  static createBook(bookData) {
-    const newBook = new Book(bookData.title, bookData.author);
-
-    Books.getBooks().push(newBook);
-    this.insertBookStructure(newBook, false);
-  }
-
-  static removeBook(div) {
-    const removeBttn = document
-      .getElementById(div.id)
-      .querySelector('.removeButton');
-    removeBttn.addEventListener('click', () => {
-      Books.getBooks().splice(div.id, 1); // eslint-disable-next-line no-use-before-define
-      CheckInput();
-      div.remove();
-    });
-  }
-
-  static insertBookStructure(book, initialDisplay) {
-    const listItem = document.createElement('li');
-    if (Books.ids % 2 === 0) {
-      listItem.classList.add('dark-bg');
-    } else {
-      listItem.classList.add('white-bg');
-    }
-    listItem.innerHTML = ` ${book.title}`
-      + ' by '
-      + `${book.author}`
-      + ' '
-      + "<button class = 'removeButton'>Remove</button><br>";
-    if (!initialDisplay) {
-      listItem.id = `${Books.getBooks().length - 1}`;
-    } else {
-      listItem.id = `${Books.ids}`;
-      Books.ids += 1;
-    }
-    document.querySelector('#booksDisplay ul').appendChild(listItem);
-    this.removeBook(listItem);
-  }
-}
-
-function localStorageAv() {
-  const test = 'test';
-  try {
-    localStorage.setItem(test, test);
-    localStorage.removeItem(test);
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
-
-function HandleInputData() {
-  const jsonData = JSON.stringify(Books.getBooks());
-  localStorage.setItem('data', jsonData);
-}
-
-const storageAvailability = localStorageAv();
-
-const CheckInput = () => {
-  if (storageAvailability) {
-    HandleInputData();
-  }
-};
 
 function hasValue(input) {
   if (input === '') {
@@ -106,7 +18,7 @@ form.addEventListener('click', (event) => {
   if (hasValue(titleValue) && hasValue(authorValue)) {
     const book = new Book(titleValue, authorValue);
     Methods.createBook(book);
-    CheckInput();
+    Methods.CheckInput();
     event.preventDefault();
   }
 });
@@ -124,4 +36,3 @@ function CheckLocalInput() {
 }
 
 CheckLocalInput();
-/* eslint max-classes-per-file: off */
